Fix all compilation warnings (146 → 0) across entire project
Header fixes (highest impact - propagate across all translation units):
- FftwArray.h: Add int32_t operator overloads alongside uint32_t for safe indexing
- DocumentSVG.h: Remove template-id from constructors (C++20 compatibility)
- Stylable.h: Comment out unused 'tag' parameter

Source file fixes:
- StreamLine.cpp: Fix shadow variable 'angle', sign/float conversions, use size_t loops
- SkeletonGrid.cpp: Fix int32_t↔uint32_t sign conversions with explicit casts
- GraphicRendering.cpp: Cast fill_color() to uint32_t, Color getters to int32_t for svg::Color
- Routing.cpp: Rename shadowing 'qEle' to 'topEle', cast id to size_t for vector::at()
- ShapeContext.cpp: Comment out unused arc parameters (rx, ry, x_axis_rotation, etc.)
- PenStroke.cpp: Add static_cast<uint32_t> for double→uint32_t config values
- Cell.cpp: Use std::size_t for loop variables, cast quantity to int32_t
- Ribbon.cpp: Use std::size_t loops, fix shadow variable 'i' → 'j', cast number to size_t
- GridIndex.cpp: Cast fill_color() to uint32_t for Color getters
- LinearGradient.cpp: Rename shadow params '_thickness1/2' → 'thickness1/2'
- PathRendering.cpp: Rename shadow 'nodes' → 'pcNodes', remove unused variable 'arr'
- Placement.cpp: Cast nets.size() to int32_t, congestion to size_t
- SpatialIndex.cpp: Comment out unused parameters 'net' and 'pc'
- NodeRendering.cpp: Comment out unused parameter 'polyConvexList'
- Context.cpp: Explicitly initialize Stylable base class in copy constructor
- HqNoise.cpp: Change sgn() to take uint32_t, remove unnecessary int32_t casts

Result: 0 warnings from project source code (was 146), all 169 tests pass

+++ LabyPath/src/Anisotrop/Placement.cpp
@@ -96,7 +96,7 @@ Routing Placement::create_route(Cell& cell) {
             if (&vertex2 != &queue.front().vertex()) {
 
                 Pin pin_target { vertex2, std::max(queue.front().thickness() / _config.decrement_factor(), _config.minimal_thickness()) };
-                Net n(queue.front(), pin_target, nets.size());
+                Net n(queue.front(), pin_target, static_cast<int32_t>(nets.size()));
 
                 if (routing.findRoute(n)) {
 
@@ -123,7 +123,7 @@ void Placement::statistics(const Arrangement_2& arr) {
 
     std::size_t total_congestion = 0;
     for (const Halfedge& he : RangeHelper::make(arr.edges_begin(), arr.edges_end())) {
-        total_congestion += he.curve().data().congestion();
+        total_congestion += static_cast<std::size_t>(he.curve().data().congestion());
     }
     std::cout << "seed  " << _config.routing().seed() << " total_congestion " << total_congestion << " number of edge " << arr.number_of_edges() << std::endl;
